GEMV cleanup

+++ tests/blas/nodes/gemv_test.py
@@ -39,9 +39,7 @@ def pure_graph(dtype, transposed):
     x = state.add_read("x")
     result = state.add_write("y")
 
-    gemv_node = blas.Gemv("gemv",
-                          dtype=dace.float32,
-                          transA=transposed)
+    gemv_node = blas.Gemv("gemv", dtype=dace.float32, transA=transposed)
 
     state.add_memlet_path(A,
                           gemv_node,
@@ -51,14 +49,12 @@ def pure_graph(dtype, transposed):
     state.add_memlet_path(x,
                           gemv_node,
                           dst_conn="_x",
-                          memlet=Memlet.simple(x,
-                                               "0:{}".format(x_size)))
+                          memlet=Memlet.simple(x, "0:{}".format(x_size)))
     y = state.add_read("y")
     state.add_memlet_path(y,
                           gemv_node,
                           dst_conn="_y",
-                          memlet=Memlet.simple(y,
-                                               "0:{}".format(y_size)))
+                          memlet=Memlet.simple(y, "0:{}".format(y_size)))
     state.add_memlet_path(gemv_node,
                           result,
                           src_conn="_y",
@@ -95,29 +91,39 @@ def intel_fpga_graph(dtype, transposed, vec_width=4):
     in_host_x = copy_in_state.add_read("x")
     in_host_y = copy_in_state.add_read("y")
 
-    sdfg.add_array("device_A", shape=[A_rows, A_cols], dtype=dtype, storage=dace.dtypes.StorageType.FPGA_Global,
+    sdfg.add_array("device_A",
+                   shape=[A_rows, A_cols],
+                   dtype=dtype,
+                   storage=dace.dtypes.StorageType.FPGA_Global,
                    transient=True)
-    sdfg.add_array("device_x", shape=[x_size], dtype=dtype, storage=dace.dtypes.StorageType.FPGA_Global,
+    sdfg.add_array("device_x",
+                   shape=[x_size],
+                   dtype=dtype,
+                   storage=dace.dtypes.StorageType.FPGA_Global,
                    transient=True)
-    sdfg.add_array("device_y", shape=[y_size], dtype=dtype, storage=dace.dtypes.StorageType.FPGA_Global,
+    sdfg.add_array("device_y",
+                   shape=[y_size],
+                   dtype=dtype,
+                   storage=dace.dtypes.StorageType.FPGA_Global,
                    transient=True)
 
     in_device_A = copy_in_state.add_write("device_A")
     in_device_x = copy_in_state.add_write("device_x")
     in_device_y = copy_in_state.add_write("device_y")
 
-    copy_in_state.add_memlet_path(
-        in_host_A, in_device_A,
-        memlet=Memlet.simple(in_host_A, "0:{}, 0:{}".format(A_rows, A_cols))
-    )
-    copy_in_state.add_memlet_path(
-        in_host_x, in_device_x,
-        memlet=Memlet.simple(in_host_x, "0:{}".format(x_size))
-    )
-    copy_in_state.add_memlet_path(
-        in_host_y, in_device_y,
-        memlet=Memlet.simple(in_host_y, "0:{}".format(y_size))
-    )
+    copy_in_state.add_memlet_path(in_host_A,
+                                  in_device_A,
+                                  memlet=Memlet.simple(
+                                      in_host_A,
+                                      "0:{}, 0:{}".format(A_rows, A_cols)))
+    copy_in_state.add_memlet_path(in_host_x,
+                                  in_device_x,
+                                  memlet=Memlet.simple(in_host_x,
+                                                       "0:{}".format(x_size)))
+    copy_in_state.add_memlet_path(in_host_y,
+                                  in_device_y,
+                                  memlet=Memlet.simple(in_host_y,
+                                                       "0:{}".format(y_size)))
 
     ###########################################################################
     # Copy data from FPGA
@@ -127,10 +133,10 @@ def intel_fpga_graph(dtype, transposed, vec_width=4):
     out_device = copy_out_state.add_read("device_y")
     out_host = copy_out_state.add_write("y")
 
-    copy_out_state.add_memlet_path(
-        out_device, out_host,
-        memlet=Memlet.simple(out_host, "0:{}".format(y_size))
-    )
+    copy_out_state.add_memlet_path(out_device,
+                                   out_host,
+                                   memlet=Memlet.simple(out_host,
+                                                        "0:{}".format(y_size)))
 
     ########################################################################
     # FPGA State
@@ -143,38 +149,41 @@ def intel_fpga_graph(dtype, transposed, vec_width=4):
     y_in = fpga_state.add_read("device_y")
     y_out = fpga_state.add_write("device_y")
 
-
-    gemv_node = blas.Gemv("gemv", dtype=dace.float32, vec_width=vec_width, transA=transposed)
+    gemv_node = blas.Gemv("gemv",
+                          dtype=dace.float32,
+                          vec_width=vec_width,
+                          transA=transposed)
     gemv_node.implementation = "IntelFPGA"
 
     fpga_state.add_memlet_path(A,
                                gemv_node,
                                dst_conn="_A",
-                               memlet=Memlet.simple(A, "0:{}, 0:{}".format(n, m)))
+                               memlet=Memlet.simple(A,
+                                                    "0:{}, 0:{}".format(n, m)))
 
     fpga_state.add_memlet_path(x,
                                gemv_node,
                                dst_conn="_x",
-                               memlet=Memlet.simple(x, "0:{}".format("{}".format(x_size))))
+                               memlet=Memlet.simple(
+                                   x, "0:{}".format("{}".format(x_size))))
     fpga_state.add_memlet_path(y_in,
                                gemv_node,
                                dst_conn="_y",
-                               memlet=Memlet.simple(y_in, "0:{}".format(y_size)))
+                               memlet=Memlet.simple(y_in,
+                                                    "0:{}".format(y_size)))
     fpga_state.add_memlet_path(gemv_node,
                                y_out,
                                src_conn="_y",
-                               memlet=Memlet.simple(y_out, "0:{}".format(y_size)))
+                               memlet=Memlet.simple(y_out,
+                                                    "0:{}".format(y_size)))
 
     ######################################
     # Interstate edges
-    sdfg.add_edge(copy_in_state, fpga_state,
-                  dace.sdfg.sdfg.InterstateEdge())
-    sdfg.add_edge(fpga_state, copy_out_state,
-                  dace.sdfg.sdfg.InterstateEdge())
+    sdfg.add_edge(copy_in_state, fpga_state, dace.sdfg.sdfg.InterstateEdge())
+    sdfg.add_edge(fpga_state, copy_out_state, dace.sdfg.sdfg.InterstateEdge())
 
     sdfg.fill_scope_connectors()
     return sdfg
-
 
 
 def fpga_row_streamd_graph(dtype, vendor, n_tile=4, m_tile=4, vec_width=1):
@@ -200,18 +209,16 @@ def fpga_row_streamd_graph(dtype, vendor, n_tile=4, m_tile=4, vec_width=1):
     if b != 0:
         test_sdfg.add_symbol(b.name, dace.float32)
 
-    test_sdfg.add_array('A', shape=[nRows*mCols/vec_width], dtype=vec_type)
-    test_sdfg.add_array('x', shape=[mCols/vec_width], dtype=vec_type)
+    test_sdfg.add_array('A', shape=[nRows * mCols / vec_width], dtype=vec_type)
+    test_sdfg.add_array('x', shape=[mCols / vec_width], dtype=vec_type)
     test_sdfg.add_array('y', shape=[nRows], dtype=single_vec_type)
     test_sdfg.add_array('res', shape=[nRows], dtype=single_vec_type)
 
-    x_stream = streaming.StreamReadVector(
-        'x',
-        mCols,
-        dace.float32,
-        veclen=vec_width,
-        repeat='{}/{}'.format(nRows, rowTile)
-    )
+    x_stream = streaming.StreamReadVector('x',
+                                          mCols,
+                                          dace.float32,
+                                          veclen=vec_width,
+                                          repeat='{}/{}'.format(nRows, rowTile))
 
     y_stream = None
     if b != 0:
@@ -222,59 +229,38 @@ def fpga_row_streamd_graph(dtype, vendor, n_tile=4, m_tile=4, vec_width=1):
             veclen=1,
         )
 
-    A_stream = streaming.StreamReadMatrixFull(
-        'A',
-        nRows,
-        mCols,
-        rowTile,
-        colTile,
-        dace.float32,
-        tileByRow=True,
-        veclen=vec_width
-    )
+    A_stream = streaming.StreamReadMatrixFull('A',
+                                              nRows,
+                                              mCols,
+                                              rowTile,
+                                              colTile,
+                                              dace.float32,
+                                              tileByRow=True,
+                                              veclen=vec_width)
 
-    res_stream = streaming.StreamWriteVector(
-        'res',
-        nRows,
-        dace.float32
-    )
+    res_stream = streaming.StreamWriteVector('res', nRows, dace.float32)
 
-    gemv_node = blas.gemv.Gemv(
-        "blas_gemv",
-        dtype=dace.float32,
-        n_tile=rowTile,
-        m_tile=colTile,
-        partial_width=partial_width,
-        n=nRows,
-        m=mCols,
-        veclen=vec_width,
-        alpha=a, beta=b,
-        streaming=True
-    )
+    gemv_node = blas.gemv.Gemv("blas_gemv",
+                               dtype=dace.float32,
+                               n_tile=rowTile,
+                               m_tile=colTile,
+                               partial_width=partial_width,
+                               n=nRows,
+                               m=mCols,
+                               veclen=vec_width,
+                               alpha=a,
+                               beta=b,
+                               streaming=True)
     gemv_node.implementation = 'fpga_row'
 
     if y_stream is not None:
         preState, postState = streaming.fpga_setup_connect_streamers(
-            test_sdfg,
-            test_state,
-            gemv_node,
-            [x_stream, y_stream, A_stream],
-            ['_x', '_y', '_A'],
-            gemv_node,
-            [res_stream],
-            ['_res']
-        )
+            test_sdfg, test_state, gemv_node, [x_stream, y_stream, A_stream],
+            ['_x', '_y', '_A'], gemv_node, [res_stream], ['_res'])
     else:
         preState, postState = streaming.fpga_setup_connect_streamers(
-            test_sdfg,
-            test_state,
-            gemv_node,
-            [x_stream, A_stream],
-            ['_x', '_A'],
-            gemv_node,
-            [res_stream],
-            ['_res']
-        )
+            test_sdfg, test_state, gemv_node, [x_stream, A_stream],
+            ['_x', '_A'], gemv_node, [res_stream], ['_res'])
 
     test_sdfg.expand_library_nodes()
 
@@ -292,7 +278,10 @@ if __name__ == "__main__":
     parser.add_argument("M", type=int, nargs="?", default=64)
     parser.add_argument("n_tile", type=int, nargs="?", default=16)
     parser.add_argument("m_tile", type=int, nargs="?", default=16)
-    parser.add_argument("alpha", type=int, nargs="?", default=1)
+    parser.add_argument("alpha",
+                        type=np.float32,
+                        nargs="?",
+                        default=np.float32(1.0))
     parser.add_argument("beta", type=int, nargs="?", default=0)
     parser.add_argument("--veclen", type=int, default=1)
     parser.add_argument('--cached', default=False, action='store_true')
@@ -303,8 +292,8 @@ if __name__ == "__main__":
     parser.add_argument("--target", dest="target", default="pure")
 
     args = parser.parse_args()
-    n = args.N
-    m = args.M
+    n = np.int32(args.N)
+    m = np.int32(args.M)
     alpha = args.alpha
     beta = args.beta
     transposed = args.transposed
@@ -313,23 +302,18 @@ if __name__ == "__main__":
     elif args.target == "intel_fpga":
         sdfg = intel_fpga_graph(dace.float32, transposed)
     elif args.target == "xilinx":
-        sdfg = fpga_row_streamd_graph(
-            dace.float32,
-            "xilinx",
-            n_tile=args.n_tile,
-            m_tile=args.m_tile,
-            vec_width=args.veclen
-        )
+        sdfg = fpga_row_streamd_graph(dace.float32,
+                                      "xilinx",
+                                      n_tile=args.n_tile,
+                                      m_tile=args.m_tile,
+                                      vec_width=args.veclen)
     else:
         print("Unsupported target")
         exit(-1)
 
-    
     if args.cached:
         print("Running from compilation cache")
         dace.config.Config.set("compiler", "use_cache", value="true")
-
-
 
     A = np.random.rand(n, m).astype(np.float32)
     x = np.random.rand(n if transposed else m).astype(np.float32)
@@ -352,7 +336,9 @@ if __name__ == "__main__":
 
     if diff >= 1e-5:
         result = res if args.target == "xilinx" else y
-        raise RuntimeError("Unexpected result returned from gemv operation ({}) with diff {} "
-                           "got:\n{}\nexpected:\n{}".format(args.target, round(diff, 6), result, ref))
+        raise RuntimeError(
+            "Unexpected result returned from gemv operation ({}) with diff {} "
+            "got:\n{}\nexpected:\n{}".format(args.target, round(diff, 6),
+                                             result, ref))
     else:
         print("Ok")
